% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aee.R
\name{aee}
\alias{aee}
\title{Ancestral Transcriptome State Inference}
\usage{
aee(x, phy, mat, select = c("descendents", "all"), CI = TRUE)
}
\arguments{
\item{x}{a vector of known expression or TF binding score values of genes, preferably log2-transformed}

\item{phy}{a phylogenetic tree in the form of object "phylo"}

\item{mat}{a matrix generated from "varMatInv" function}

\item{select}{indicate if descendents of the node or all tips should be used}

\item{CI}{a logical specifying whether to return the 95% confidence intervals
of the estimated ancestral transcriptome value.}
}
\value{
returns a list containing estimated ancestral transcriptome states
as well as other requested parameters
}
\description{
This function is to inference ancestral transcriptome state
and related statistical uncertainty based on RNA-seq expression data
or ChIP-seq TF-binding data.
}
