% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aee.tf.R
\name{aeeBS}
\alias{aeeBS}
\title{Ancestral TF-binding State Inference}
\usage{
aeeBS(x, phy, mat, select = c("descendents", "all"), CI = TRUE)
}
\arguments{
\item{x}{a vector of known TF binding score values of a gene, preferably log2-transformed TF binding scores}

\item{phy}{a phylogenetic tree in the form of object "phylo"}

\item{mat}{a matrix generated from "varMatInv" function}

\item{select}{indicate if descendents of the node or all tips should be used}

\item{CI}{a logical specifying whether to return the 95% confidence intervals
of the estimated ancestral TF binding scores}
}
\value{
returns a list containing estimated ancestral TF binding states
as well as other requested parameters
}
\description{
This function is to inference ancestral TF binding state and related statistical uncertainty
}
