% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnceTran.R
\docType{package}
\name{AnceTran}
\alias{AnceTran}
\alias{AnceTran-package}
\title{AnceTran: the package for phylogenetic analysis of TF-binding data}
\description{
\code{AnceTran} is an R package that performs TF-binding phylogenetic analysis
from ChIP-seq data, including optimized input formatting, normalization
and pair-wise distance evaluation, ancestral TF-binding state inference.
}
\author{
Jingw Yang (jingwyang.fudan@gmail.com).
}
