% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.sou.tf.R
\name{distances}
\alias{distances}
\alias{TFdist.sou}
\title{Function for estimating pair-wise TF binding distances based on sOU model}
\usage{
TFdist.sou(bsMat = NULL)
}
\arguments{
\item{bsMat}{a TF-binding score matrix: column corresponds to binding score value;
row corresponds to othologous genes}
}
\value{
A distance matrix shows the sOU distance for each paried specis.
}
\description{
sOU method to estimate pair-wise TF binding distances
}
