% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF.objects.R
\docType{data}
\name{TF.objects}
\alias{TF.objects}
\title{taxaTF object created from HNF4A binding data for 4 mice species}
\format{a \code{taxaTF} object, contains 4 \code{taxonExp}:
with raw binding score values on 22698 genes.}
\description{
large \code{taxaTF} object created from
HNF4A-binding data of four mice species
}
\examples{
data(TF.objects)
TF.objects
TF.objects[[1]]
}
\references{
Wong,ES. et al. (2015) Decoupling of evolutionary changes
in transcription factor binding and gene expression in mammals. Genome Res 25: 167-178.
}
