% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFtab.R
\name{TFtab}
\alias{TFtab}
\title{TF-binding score table extracted from a \code{taxaTF} class}
\usage{
TFtab(objects = NULL, taxa = "all", tf = "all", rowindex = NULL,
  filtering = TRUE, normalize = TRUE, logrithm = TRUE)
}
\arguments{
\item{objects}{a vector of objects of class \code{taxonTF} or an object of class \code{taxaTF}}

\item{taxa}{one single character or a vector of characters specifying main taxa to be included in
the TF binding score table. taxa usually corresponds to species tissues or cell line types.
If one single character "all" is given,
all the taxa included in the \code{taxaTF} will be matched and included ("all" by default).}

\item{tf}{one single character or a vector of characters specifying transcription factor(s) to be included in
the TF binding score table.
If one single character "all" is given,
all the transcription factor included in the \code{taxaTF} will be matched and included ("all" by default).}

\item{rowindex}{a vector of numbers corresponded to indices of selecting rows}

\item{filtering}{a logical specifying whether to exclude genes with binding score equal to 0 in all taxaon and tfs.}

\item{normalize}{a logical specifying whether to perform quantile normalization on sample-specific biases.}

\item{logrithm}{a logical specifying whether to apply TF binding score log2 tranformation (TRUE by default).}
}
\value{
An TF binding score table: column corresponds to median binding score value of all biological samples
within one taxa_TF group; row corresponds to othologous genes
}
\description{
Generate an TF binding score table from a \code{taxaTF} class
}
