% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnceTran.R
\docType{package}
\name{AnceTran}
\alias{AnceTran}
\alias{AnceTran-package}
\title{AnceTran: the package for comparative phylogenetic analysis of transcriptome  data}
\description{
\code{AnceTran} is an R package that performs transcriptome phylogenetic analysis
based on RNA-seq expression data and ChIP-seq TF-binding data,
including optimized input formatting, normalization
and pair-wise distance evaluation, character tree building and
ancestral transcriptome state inference.
}
\author{
Jingw Yang (jingwyang.fudan@gmail.com);  Hang Ruan (hang.ruan@hotmail.com).
}
