% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpDistances.R, R/TFDistances.R
\name{distances}
\alias{distances}
\alias{Expdist.nbdln}
\alias{Expdist.u}
\alias{Expdist.jsd}
\alias{Expdist.pea}
\alias{Expdist.spe}
\alias{Expdist.euc}
\alias{Expdist.cos}
\alias{Expdist.tani}
\alias{Expdist.jac}
\alias{Expdist.ced}
\alias{Expdist.sou}
\alias{distances}
\alias{TFdist.sou}
\title{Internal functions for estimating pair-wise transcriptome distances}
\usage{
Expdist.nbdln(reads.count = NULL, gene_length = NULL, omega = NULL)

Expdist.u(reads.count = NULL, gene_length = NULL)

Expdist.jsd(expMat = NULL)

Expdist.pea(expMat = NULL)

Expdist.spe(expMat = NULL)

Expdist.euc(expMat = NULL)

Expdist.cos(expMat = NULL)

Expdist.tani(expMat = NULL)

Expdist.jac(expMat = NULL)

Expdist.ced(expMat = NULL)

Expdist.sou(expMat = NULL)

TFdist.sou(bsMat = NULL)
}
\arguments{
\item{expMat}{a expression value matrix: colum correspnds to expression values;
row correspnds to othologous genes}

\item{bsMat}{a TF-binding score matrix: column corresponds to binding score value;
row corresponds to othologous genes}
}
\value{
A distance matrix shows the transcriptome distance for each paried species.
}
\description{
Several published methods to estimate pair-wise transcriptome distances.

Distance methods used to analyze expression values include Pearson distance,
Euclidean distance, angular-cosine distance, the square root of
Jensen-Shannon divergence, the Brownian distance, and so on.

Stationary OU method is applied to estimate pair-wise TF binding distances
}
\references{
Chen H, He X. 2016. The Convergent Cancer Evolution toward a Single Cellular Destination.
Mol Biol Evol 33:4-12

Gu X, Su Z. 2007. Tissue-driven hypothesis of genomic evolution and sequence-expression correlations.
Proc Natl Acad Sci USA 104:2779-2784.

Pereira V, Waxman D, Eyre-Walker A. 2009. A problem with the correlation coefficient as a measure of gene expression divergence.
Genetics 183:1597-1600.

Sudmant PH, Alexis MS, Burge CB. 2015. Meta-analysis of RNA-seq expression data across species, tissues and studies.
Genome Biol 16:287.

Distance based on negative bio distribution and log normal model
}
