% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varMatInv.R
\name{varMatInv}
\alias{varMatInv}
\title{Generate an inversed variance-covariance matrix from transcriptome profiles across species}
\usage{
varMatInv(dismat, tran_table, phy)
}
\arguments{
\item{dismat}{a distance matrix for paired transcriptome profiles}

\item{tran_table}{transcriptome value matrix extracted from objects of class
\code{taxaTF} or class \code{taxaExp}}

\item{phy}{an rooted transcriptome character tree}
}
\value{
returns an inversed variance-covariance matrix
}
\description{
This function generate an inversed variance-covariance matrix from expression
or TF-binding profiles of one-to-one orthologous genes across species
}
